make new module SimulatorFactory
