+++ src/Simulator/ISimulator.ts
@@ -5,40 +5,6 @@ import {
     Plot,
     Metrics,
 } from "../simularium/types.js";
-import { ClientSimulator } from "./ClientSimulator.js";
-import { LocalFileSimulator } from "./LocalFileSimulator.js";
-import { RemoteSimulator } from "./RemoteSimulator.js";
-
-import {
-    SimulatorParams,
-    RemoteSimulatorParams,
-    ClientSimulatorParams,
-    LocalFileSimulatorParams,
-} from "./types.js";
-
-export const getSimulatorClassFromParams = (params?: SimulatorParams) => {
-    if (!params || !params.fileName) {
-        return { simulatorClass: null, typedParams: null };
-    }
-    if ("netConnectionSettings" in params) {
-        return {
-            simulatorClass: RemoteSimulator,
-            typedParams: params as RemoteSimulatorParams,
-        };
-    } else if ("clientSimulatorImpl" in params) {
-        return {
-            simulatorClass: ClientSimulator,
-            typedParams: params as ClientSimulatorParams,
-        };
-    } else if ("simulariumFile" in params) {
-        return {
-            simulatorClass: LocalFileSimulator,
-            typedParams: params as LocalFileSimulatorParams,
-        };
-    } else {
-        return { simulatorClass: null, typedParams: null };
-    }
-};
 
 /**
 From the caller's perspective, this interface is a contract for a
